# wallet-bff-ws-3/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "R2PS REST API",
    "description": "REST API Level 3 (HATEOAS) for Remote to Physical Signing service.\n\nREST API Nivå 3 (HATEOAS) för tjänsten Remote to Physical Signing (R2PS). Tillhandahålls av DIGG – Myndigheten för digital förvaltning.\n\n## Known Issues / Kända problem\n\n- Synchronous mode may time out for large payloads. Use asynchronous mode for requests expected to take longer than 5 seconds.\n- Synkront läge kan få timeout vid stora nyttolaster. Använd asynkront läge för förfrågningar som förväntas ta längre än 5 sekunder.",
    "termsOfService": "https://www.digg.se/digitala-tjanster/e-legitimering/tillitsramverk",
    "contact": {
      "name": "DIGG – Myndigheten för digital förvaltning",
      "url": "https://www.digg.se",
      "email": "info@digg.se"
    },
    "license": {
      "name": "EUPL-1.2",
      "url": "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
    },
    "version": "0.1.0",
    "x-service-level": {
      "description": "This API is currently in alpha. No SLA guarantees are provided. The service targets 99% availability during business hours (CET/CEST).",
      "beskrivning": "Detta API är för närvarande i alfa. Inga SLA-garantier ges. Tjänsten siktar på 99% tillgänglighet under kontorstid (CET/CEST)."
    },
    "x-known-issues": [
      {
        "description": "Synchronous mode may time out for large payloads. Use asynchronous mode for requests expected to take longer than 5 seconds.",
        "beskrivning": "Synkront läge kan få timeout vid stora nyttolaster. Använd asynkront läge för förfrågningar som förväntas ta längre än 5 sekunder."
      }
    ]
  },
  "paths": {
    "/": {
      "post": {
        "tags": [
          "R2PS API"
        ],
        "summary": "POST / - Main service request endpoint",
        "description": "Accepts a service request and either returns immediately with a pending status\n(async mode) or waits for the result (sync mode).",
        "operationId": "submit_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Request completed successfully (sync mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncResponseDto"
                }
              }
            }
          },
          "202": {
            "description": "Request accepted and pending (async mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Device not found"
          },
          "408": {
            "description": "Request timeout (sync mode)"
          }
        }
      }
    },
    "/api-info": {
      "get": {
        "tags": [
          "API Metadata"
        ],
        "summary": "GET /api-info - API Information endpoint",
        "description": "Returns API metadata as required by Swedish REST API Profile (VER.06, VER.07).",
        "operationId": "api_info",
        "responses": {
          "200": {
            "description": "API information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/device-states": {
      "post": {
        "tags": [
          "R2PS API"
        ],
        "summary": "POST /device-states - Device state initialization endpoint",
        "description": "Initialize a new device with a public key and obtain a client ID.\n**WARNING**: Contains dev-only features (overwrite, clientId) that should be removed for production.",
        "operationId": "new_state",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewStateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Device initialized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewStateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "408": {
            "description": "Timeout waiting for HSM response"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health check",
        "description": "Returns HTTP 200 OK if the service is running and ready to accept requests.\nThis endpoint can be used by load balancers, orchestrators, and monitoring\nsystems to verify that the R2PS service is operational. It performs no\ndependency checks (Redis, Kafka) — it only confirms the HTTP server is responding.",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Service is running and ready to accept requests"
          }
        }
      }
    },
    "/requests/{correlationId}": {
      "get": {
        "tags": [
          "R2PS API"
        ],
        "summary": "GET /requests/{correlationId} - Response polling endpoint",
        "description": "Poll for the result of an async request using the correlation ID.",
        "operationId": "poll_task",
        "parameters": [
          {
            "name": "correlationId",
            "in": "path",
            "description": "Correlation ID of the pending request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncResponseDto"
                }
              }
            }
          },
          "202": {
            "description": "Request still pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiInfoDto": {
        "type": "object",
        "description": "API Information response (VER.06, VER.07).\nRequired by Swedish REST API Profile v1.2.0.",
        "required": [
          "apiName",
          "apiVersion",
          "apiReleased",
          "apiDocumentation",
          "apiStatus"
        ],
        "properties": {
          "apiDocumentation": {
            "type": "string",
            "description": "URL to API documentation"
          },
          "apiName": {
            "type": "string",
            "description": "Name of the API"
          },
          "apiReleased": {
            "type": "string",
            "description": "Release date (RFC 3339 format)"
          },
          "apiStatus": {
            "type": "string",
            "description": "Lifecycle status: alpha, beta, active, deprecated, retired, decommissioned"
          },
          "apiVersion": {
            "type": "string",
            "description": "Full version number (MAJOR.MINOR.PATCH)"
          }
        },
        "example": {
          "apiName": "r2ps-api",
          "apiVersion": "0.1.0",
          "apiReleased": "2024-01-01",
          "apiDocumentation": "https://api.digg.se/r2ps-api/v1/",
          "apiStatus": "alpha"
        }
      },
      "AsyncResponseDto": {
        "type": "object",
        "description": "Generic async response wrapper with HATEOAS support.",
        "required": [
          "correlationId",
          "status",
          "_links"
        ],
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid",
            "description": "Correlation ID for tracking the request"
          },
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AsyncResponseError"
              }
            ],
            "nullable": true
          },
          "result": {
            "allOf": [
              {
                "$ref": "#/components/schemas/T"
              }
            ],
            "nullable": true
          },
          "resultUrl": {
            "type": "string",
            "description": "URL for polling the result",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AsyncResponseStatus"
          }
        },
        "example": {
          "correlationId": "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "status": "PENDING",
          "resultUrl": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
          "_links": {
            "self": {
              "href": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
              "rel": "self",
              "method": "GET"
            },
            "poll": {
              "href": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
              "rel": "poll",
              "method": "GET"
            }
          }
        }
      },
      "AsyncResponseError": {
        "type": "object",
        "description": "Error information in async response.",
        "required": [
          "message",
          "httpStatus"
        ],
        "properties": {
          "httpStatus": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "HSM worker returned an error",
          "httpStatus": 500
        }
      },
      "AsyncResponseStatus": {
        "type": "string",
        "description": "Async response status.",
        "enum": [
          "PENDING",
          "COMPLETE",
          "ERROR"
        ]
      },
      "BffRequest": {
        "type": "object",
        "description": "Request from BFF to the REST API.",
        "required": [
          "clientId",
          "outerRequestJws"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Device/wallet identifier"
          },
          "outerRequestJws": {
            "type": "string",
            "description": "JWS-signed service request envelope"
          }
        },
        "example": {
          "clientId": "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "outerRequestJws": "eyJhbGciOiJFUzI1NiJ9.eyJzdWIiOiIxMjM0NTY3ODkwIn0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        }
      },
      "EcPublicJwk": {
        "type": "object",
        "description": "EC Public Key in JWK format.\n\nThe `kid` field is required and identifies this key within the device's\nkey set (RFC 7517). The HSM worker uses it to look up the correct\npublic key for JWS signature verification.",
        "required": [
          "kty",
          "crv",
          "x",
          "y",
          "kid"
        ],
        "properties": {
          "crv": {
            "type": "string"
          },
          "kid": {
            "type": "string",
            "description": "Key identifier (RFC 7517) — required, must not be empty"
          },
          "kty": {
            "type": "string"
          },
          "x": {
            "type": "string"
          },
          "y": {
            "type": "string"
          }
        },
        "example": {
          "kty": "EC",
          "crv": "P-256",
          "x": "f83OJ3D2xF1Bg8vub9tLe1gHMzV76e8Tus9uPHvRVEU",
          "y": "x_FEzRu9m36HLN_tue659LNpXW6pCyStikYjKIWI5a0",
          "kid": "device-key-1"
        }
      },
      "Link": {
        "type": "object",
        "description": "HATEOAS link following REST Level 3 principles (HYP.11-HYP.17).\nComplies with Swedish REST API Profile v1.2.0.",
        "required": [
          "href",
          "method"
        ],
        "properties": {
          "href": {
            "type": "string",
            "description": "URI of the related resource (MUST be absolute URL - HYP.07, HYP.12)"
          },
          "method": {
            "type": "string",
            "description": "HTTP method for the link - MUST be present even for GET (HYP.17)"
          },
          "rel": {
            "type": "string",
            "description": "Relation type (e.g., \"self\", \"next\", \"poll\", \"cancel\") - MUST be present (HYP.15)",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Media type hint (optional)",
            "nullable": true
          }
        },
        "example": {
          "href": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
          "rel": "self",
          "method": "GET"
        }
      },
      "Links": {
        "type": "object",
        "description": "HATEOAS links collection. The `self` link is mandatory per REST Level 3 (HYP.11). Additional links (e.g., poll, cancel) may be present.",
        "required": [
          "self"
        ],
        "properties": {
          "self": {
            "$ref": "#/components/schemas/Link"
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        },
        "example": {
          "self": {
            "href": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
            "rel": "self",
            "method": "GET"
          },
          "poll": {
            "href": "https://api.digg.se/r2ps-api/v1/requests/d290f1ee-6c54-4b01-90e6-d701748f0851",
            "rel": "poll",
            "method": "GET"
          }
        }
      },
      "NewStateRequestDto": {
        "type": "object",
        "description": "Request to initialize new device state.",
        "required": [
          "publicKey",
          "ttl"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Optional client ID (dev-only, should be removed in production)",
            "nullable": true
          },
          "overwrite": {
            "type": "boolean",
            "description": "Dev-only flag to overwrite existing state",
            "nullable": true
          },
          "publicKey": {
            "$ref": "#/components/schemas/EcPublicJwk"
          },
          "ttl": {
            "type": "string",
            "description": "TTL in ISO 8601 duration format (e.g., \"P30D\", \"PT1H\")"
          }
        },
        "example": {
          "publicKey": {
            "kty": "EC",
            "crv": "P-256",
            "x": "f83OJ3D2xF1Bg8vub9tLe1gHMzV76e8Tus9uPHvRVEU",
            "y": "x_FEzRu9m36HLN_tue659LNpXW6pCyStikYjKIWI5a0",
            "kid": "device-key-1"
          },
          "ttl": "P30D"
        }
      },
      "NewStateResponseDto": {
        "type": "object",
        "description": "Response from state initialization.",
        "required": [
          "status",
          "clientId",
          "devAuthorizationCode",
          "_links"
        ],
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "clientId": {
            "type": "string"
          },
          "devAuthorizationCode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "example": {
          "status": "OK",
          "clientId": "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "devAuthorizationCode": "abc123def456",
          "_links": {
            "self": {
              "href": "https://api.digg.se/r2ps-api/v1/device-states",
              "rel": "self",
              "method": "GET"
            },
            "submit-request": {
              "href": "https://api.digg.se/r2ps-api/v1",
              "rel": "submit-request",
              "method": "POST",
              "type": "application/json"
            }
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 9457 Problem Details for HTTP APIs (FEL.01, FEL.02).\nComplies with Swedish REST API Profile v1.2.0.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence"
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type"
          },
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type"
          }
        },
        "example": {
          "type": "https://api.digg.se/r2ps/v1/problems/invalid-request",
          "title": "Invalid Request",
          "status": 400,
          "detail": "The request body is malformed or missing required fields"
        }
      }
    }
  },
  "tags": [
    {
      "name": "R2PS API",
      "description": "Remote to Physical Signing API endpoints"
    },
    {
      "name": "Health",
      "description": "Health check endpoints"
    }
  ]
}